Revert "run eslint, check CI"
