--- .eslintrc.js
@@ -1,8 +1,0 @@
-module.exports = {
-  extends: [
-    'plugin:vue/vue3-recommended',
-  ],
-  rules: {
-    // override/add rules settings here:
-  }
-}
