--- .github/workflows/reviewdog.yml
@@ -1,13 +1,0 @@
-name: review-dog 
-
-on:
-  pull_request:
-  push:
-    branches: [main]
-
-jobs:
-  eslint:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-      - uses: reviewdog/action-eslint@v1
