--- .pre-commit-config.yaml
@@ -1,10 +1,0 @@
-repos:
-  - repo: https://github.com/pre-commit/mirrors-eslint
-    rev: 'v7.18.0'
-    hooks:
-      - id: eslint
-        types: [javascript]
-        types_or: [javascript, jsx, ts, tsx, vue]
-        additional_dependencies:
-          - eslint@latest
-          - eslint-plugin-vue@v7.5.0
