--- .tool-versions
@@ -1,2 +1,0 @@
-nodejs 19.5.0
-eslint 8.40.0
