+++ frontend/package.json
@@ -17,8 +17,6 @@
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "^4.2.1",
-    "eslint": "^8.40.0",
-    "eslint-plugin-vue": "^9.13.0",
     "vite": "^4.3.4"
   }
 }
