+++ frontend/src/assets/Dunder-Mifflin-Symbol.png
