+++ frontend/src/assets/plane.png
