+++ frontend/src/components/WelcomeItem.vue
@@ -1,13 +1,13 @@
 <template>
   <div class="item">
     <i>
-      <slot name="icon" />
+      <slot name="icon"></slot>
     </i>
     <div class="details">
       <h3>
-        <slot name="heading" />
+        <slot name="heading"></slot>
       </h3>
-      <slot />
+      <slot></slot>
     </div>
   </div>
 </template>
