+++ frontend/src/plugins/vuetify.js
