+++ frontend/src/views/ForbiddenView.vue
@@ -1,0 +1,3 @@
+<template>
+    <h1>403 Forbidden</h1>
+</template>
