+++ frontend/src/views/ProjectsView.vue
@@ -20,8 +20,8 @@ const user = computed(() => userStore.user)
 </script>
 
 <template>
-  <div class="main-wrapper">
-    <!-- <v-container> -->
+<div class="main-wrapper">
+  <!-- <v-container> -->
     <!-- header -->
     <v-row class="mt-5">
       <h1>Hello, {{ user.firstName }}</h1>
@@ -32,44 +32,33 @@ const user = computed(() => userStore.user)
     <!-- end of header -->
     <!-- projects section -->
     <v-row>
-      <template
-        v-for="project in projects"
-        :key="project.id"
-      >
-        <v-col
-          xs="12"
-          sm="4"
-          md="3"
-          class="mb-2"
-        >
-          <project-card
-            :id="project.id"
-            :name="project.name"
-            :progress="(done/all)*100"
-          />
-        </v-col>
-      </template>
-      <v-col
-        xs="12"
-        sm="4"
-        md="3"
-      >
-        <v-card
-          class="d-flex align-center justify-center add-project"
-          height="200"
-          width="200"
-        >
-          <div class="text-center">
-            <v-icon
-              icon="mdi-folder-plus"
-              size="x-large"
-            />
-            <v-card-text class="project-name">
-              Create New Project
-            </v-card-text>
-          </div>
-        </v-card>
-      </v-col>
+  <template v-for="project in projects" :key="project.id">
+    <v-col
+      xs="12"
+      sm="4"
+      md="3"
+      class="mb-2"
+    >
+      <project-card
+        :id="project.id"
+        :name="project.name"
+        :progress="(done/all)*100"
+      />
+    </v-col>
+  </template>
+    <v-col
+      xs="12"
+      sm="4"
+      md="3"
+    >
+      <v-card class="d-flex align-center justify-center add-project" height="200" width="200">
+        <div class="text-center">
+          <v-icon icon="mdi-folder-plus" size="x-large"></v-icon>
+          <v-card-text class="project-name">Create New Project</v-card-text>
+        </div>
+      </v-card>
+    </v-col>
+    
     </v-row>
     <!-- end of projects section -->
   <!-- </v-container> -->
