--- src/views/ForbiddenView.vue
@@ -1,3 +1,0 @@
-<template>
-  <h1>403 Forbidden</h1>
-</template>
